Unshim ExclusiveModeGpuDiscoveryPlugin (#3590)
* Unshim ExclusiveModeGpuDiscoveryPlugin

As one of the publicly documented classes loaded by Spark outside the
plugin control, ExclusiveModeGpuDiscoveryPlugin may not reside in the
Parallel World locations. Binary dedupe shows
ExclusiveModeGpuDiscoveryPlugin is bitwise-identical across all shims


* Delay ExclusiveModeGpuDiscoveryPlugin entry point

via ShimLoader


* Fix scalastyle
