+++ sql-plugin/src/main/scala/com/nvidia/spark/ExclusiveModeGpuDiscoveryPlugin.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, NVIDIA CORPORATION.
+ * Copyright (c) 2020-2021, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,14 +18,10 @@ package com.nvidia.spark
 
 import java.util.Optional
 
-import scala.collection.mutable.ArrayBuffer
-
-import ai.rapids.cudf.Cuda
-import com.nvidia.spark.rapids.GpuDeviceManager
+import com.nvidia.spark.rapids.ShimLoader
 
 import org.apache.spark.SparkConf
 import org.apache.spark.api.resource.ResourceDiscoveryPlugin
-import org.apache.spark.internal.Logging
 import org.apache.spark.resource.{ResourceInformation, ResourceRequest}
 
 /**
@@ -39,44 +35,12 @@ import org.apache.spark.resource.{ResourceInformation, ResourceRequest}
  *  This plugin can be activated in spark with the configuration:
  *  `--conf spark.resources.discoveryPlugin=com.nvidia.spark.ExclusiveModeGpuDiscoveryPlugin`
  */
-class ExclusiveModeGpuDiscoveryPlugin extends ResourceDiscoveryPlugin with Logging {
+class ExclusiveModeGpuDiscoveryPlugin extends ResourceDiscoveryPlugin with Proxy {
   override def discoverResource(
-      request: ResourceRequest,
-      sparkconf: SparkConf): Optional[ResourceInformation] = {
+    request: ResourceRequest,
+    sparkConf: SparkConf
+  ): Optional[ResourceInformation] = self.discoverResource(request, sparkConf)
 
-    val resourceName = request.id.resourceName
-    if (!resourceName.equals("gpu")) {
-      logInfo("ExclusiveModeGpuDiscoveryPlugin only handles gpu allocations, " +
-        s"skipping $resourceName")
-      return Optional.empty()
-    }
-    val ngpusRequested = request.amount
-    val deviceCount: Int = Cuda.getDeviceCount
-    logInfo(s"Running ExclusiveModeGpuDiscoveryPlugin to acquire $ngpusRequested GPU(s), " +
-      s"host has $deviceCount GPU(s)")
-    // loop multiple times to see if a GPU was released or something unexpected happened that
-    // we couldn't acquire on first try
-    var numRetries = 2
-    val allocatedAddrs = ArrayBuffer[String]()
-    val addrsToTry = ArrayBuffer.empty ++= (0 until deviceCount)
-    while (numRetries > 0 && allocatedAddrs.size < ngpusRequested && addrsToTry.nonEmpty) {
-      var addrLoc = 0
-      val allAddrs = addrsToTry.size
-      while (addrLoc < allAddrs && allocatedAddrs.size < ngpusRequested) {
-        val addr = addrsToTry(addrLoc)
-        if (GpuDeviceManager.tryToSetGpuDeviceAndAcquire(addr)) {
-          allocatedAddrs += addr.toString
-        }
-        addrLoc += 1
-      }
-      addrsToTry --= allocatedAddrs.map(_.toInt)
-      numRetries -= 1
-    }
-    if (allocatedAddrs.size < ngpusRequested) {
-      // log warning here, Spark will throw exception if we return not enough
-      logWarning(s"ExclusiveModeGpuDiscoveryPlugin did not find enough gpus, " +
-        s"requested: $ngpusRequested found: ${allocatedAddrs.size}")
-    }
-    Optional.of(new ResourceInformation("gpu", allocatedAddrs.toArray))
-  }
+  override lazy val self: ResourceDiscoveryPlugin =
+    ShimLoader.newInternalExclusiveModeGpuDiscoveryPlugin()
 }
